Feat:SuperBuilder 구현 Refactor:Builder 코드 수정
